% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GKM-package.R
\name{cluster}
\alias{cluster}
\title{A cluster object}
\description{
A cluster object is a list containing the following:
\describe{
     \item{\code{"k"}}{The number of clusters.}
     \item{\code{"centres"}}{A matrix where each row contains the optimised cluster centres.}
     \item{\code{"neighbourhood"}}{A matrix specifying the neighbourhood of the cluster centres.}
     \item{\code{"kernel"}}{A matrix specifying the kernel of the cluster centres.}
     \item{\code{"previous"}}{A vector indicating the cluster where each data-point was assigned.}
     \item{\code{"iterations"}}{The number of iterations before convergence.}
     \item{\code{"minmax"}}{A matrix containing the smallest and largest values found along each dimension.}
}
}
