% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_cluster.R
\name{greedy_cluster}
\alias{greedy_cluster}
\title{Greedy K-means Clustering}
\usage{
greedy_cluster(x, k, control = list())
}
\arguments{
\item{x}{Numeric matrix: A matrix with rows and columns corresponding to observations and features, respectively. NB: if \code{x} is not a matrix, the function will try to cast it as a matrix.}

\item{k}{Integer, integer vector, or adjacency list: The number of clusters, the grid layout, or the adjacency matrix of the clusters, respectively.}

\item{control}{List: A list of control arguments, for more details see \link{control_greedy_cluster}.}
}
\value{
An object of class \link{cluster}.
}
\description{
Regularised K-means clustering, allowing for greedy re-allocation of data-points when updating clusters, by exploiting the pre-defined neighbourhood structure.
}
\details{
If the argument \code{k} is given as an adjacency list, then the elements will be taken as the number of clusters and the argument \code{neighbours}, set in the control-object, will be ignored.
}
