% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_cluster.R
\name{control_greedy_cluster}
\alias{control_greedy_cluster}
\title{Control arguments for the \link{greedy_cluster}-function}
\usage{
control_greedy_cluster(
  neighbours = 2L,
  neighbours_type = "regular",
  neighbours_connected = TRUE,
  rng_start = FALSE,
  greedy = 3L,
  greedy_end = FALSE,
  competitive = TRUE,
  competitive_release = 5,
  max_iteration = c(100, 2),
  learning_range = c(1, 0.1, 3),
  seed = sample(1e+06, 1),
  tolerance = (.Machine$double.eps)^(1/4),
  include_data = TRUE,
  trace = 0L
)
}
\arguments{
\item{neighbours}{Integer: The number of neighbours of each cluster.}

\item{neighbours_type}{String: The type of adjacency matrix constructed from \code{neighbours}.}

\item{neighbours_connected}{Boolean: Should the neighbourhood graph be connected?}

\item{rng_start}{Boolean: Should the clusters be initialised entirely at random, or using kmeans++?}

\item{greedy}{Integer: The number of iterations between best-match assignments.}

\item{greedy_end}{Boolean: Should the final assignment be greedy?}

\item{competitive}{Boolean: Should the clusters be allowed to compete during the training process?}

\item{competitive_release}{Integer: The number of iterations before competitive learning is stopped (only used if \code{competitive = TRUE}).}

\item{max_iteration}{Numeric vector: A tuple (the maximum number of iterations, the maximum number of iterations below the tolerance).}

\item{learning_range}{Numeric vector: A triple (start learning range, end learning range, number iterations to go from start to end range).}

\item{seed}{Numeric: A seed passed to c++.}

\item{tolerance}{Numeric: A tolerance on the absolute change in centres between iterations; used for early stopping.}

\item{include_data}{Boolean: Should the data be included in the return object?}

\item{trace}{Integer: Prints tracing information on the progress of the algorithm every \code{trace} number of iterations.}
}
\value{
A list of control arguments.
}
\description{
Creates a list of default control arguments used by the \link{greedy_cluster}-function.
}
\details{
The argument \code{competitive_release} is technically not necessary as it can be controlled by setting an extremely small learning range. However, when the learning range is low (to the point where the optimisation is not competitive) the performance computational complexity of the algorithm can be improved by not accounting for neighbours.
}
\seealso{
Other greedy cluster:
\code{\link[=greedy_cluster]{greedy_cluster()}},
\code{\link[=predict.cluster]{predict.cluster()}}
}
\concept{greedy cluster}
