% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GKM.R
\name{control_greedy_cluster}
\alias{control_greedy_cluster}
\title{Control arguments for the \link{greedy_cluster} function}
\usage{
control_greedy_cluster(
  nd = 2L,
  rngstart = FALSE,
  greedy = TRUE,
  greedy_reset = 10,
  greedy_end = FALSE,
  competitive = FALSE,
  competitive_release = NULL,
  max_iteration = c(100, 10),
  learning_range = c(1, 0.1, 5),
  seed = sample(1e+06, 1),
  tolerance = (.Machine$double.eps)^(1/4),
  trace = 0L
)
}
\arguments{
\item{nd}{Integer: The maximum distance, measured along the grid, for two clusters to be considered neighbours.}

\item{rngstart}{Boolean: Should the clusters be initialised entirely at random, or using kmeans++?}

\item{greedy}{Boolean: Should the assignment be made greedily using the neighbourhood structure?}

\item{greedy_reset}{Integer: The number of iterations between non-greedy allocations (only used if \code{greedy == TRUE}).}

\item{greedy_end}{Boolean: Should the final allocation be made greedily?}

\item{competitive}{Boolean: Should the clusters be allowed to compete during the training process?}

\item{competitive_release}{Integer: The number of iterations before competitive learning is stopped (only used if \code{competitive == TRUE}).}

\item{max_iteration}{Numeric vector: A tuple (the maximum number of iterations, the maximum number of iterations below the tolerance).}

\item{learning_range}{Numeric vector: A triple (start learning range, end learning range, number iterations to go from start to end range).}

\item{seed}{Numeric: A seed passed to c++.}

\item{tolerance}{Numeric: A tolerance on the absolute change in centres between iterations; used for early stopping.}

\item{trace}{Integer: Prints tracing information on the progress of the algorithm every \code{trace} number of iterations.}

\item{batch}{Boolean: Should the batch version of the algorithm be used?}

\item{learning_rate}{Numeric vector: A triple (start learning rate, end learning rate, number iterations to go from start to end rate).}
}
\value{
A list of control arguments.
}
\description{
Creates a list of default control arguments used by the \link{greedysom} function.
}
\details{
The argument \code{competitive_release} is technically not necessary as it can be controlled by setting an extremely small learning range. However, when the learning range is low (to the point where the optimisation is not competitive) the performance computational complexity of the algorithm can be improved by not accounting for neighbours.
}
