% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GKM-package.R
\docType{package}
\name{GKM-package}
\alias{GKM}
\alias{GKM-package}
\title{GKM: Greedy K-means Clustering}
\description{
Implements greedy re-allocation for K-means clustering.

Implements greedy re-allocation for K-means clustering.
}
\author{
\strong{Maintainer}: Søren B. Vilsen \email{svilsen@mp.aau.dk}

Authors:
\itemize{
  \item Søren B. Vilsen \email{svilsen@mp.aau.dk}
}


Søren B. Vilsen <svilsen@mp.aau.dk>
}
\keyword{internal}
