% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_cluster.R
\name{greedy_cluster}
\alias{greedy_cluster}
\title{Greedy K-means Clustering}
\usage{
greedy_cluster(x, k, control = list())
}
\arguments{
\item{x}{Numeric matrix: A matrix with rows and columns corresponding to observations and features, respectively. NB: if \code{x} is not a matrix, the function will try to cast it as a matrix.}

\item{k}{Integer, integer vector, or adjacency list: The number of clusters, the grid layout, or the adjacency matrix of the clusters, respectively.}

\item{control}{List: A list of control arguments, for more details see \link{control_greedy_cluster}.}
}
\value{
An object of class \link{cluster}.
}
\description{
Regularised K-means clustering, allowing for greedy re-assignment of data-points when updating clusters, by exploiting the pre-defined neighbourhood structure.
}
\details{
If the argument \code{k} is given as an adjacency list, then the number of clusters will be set as the size of the list. Furthermore, the argument \code{neighbours}, set in \link{control_greedy_cluster}, will be ignored.
}
\examples{
# Data
x <- iris[, -5]
cl <- iris[, 5] |> as.numeric()

# Greedy clustering using regular graph
grd_reg <- greedy_cluster(
 x,
 k = 3,
 control = list(
    rng_start = TRUE,
    greedy = 2L,
    neighbours = 1L,
    neighbours_type = "regular",
    seed = 123456
 )
)

plot(x[, 3:4], col = cl + 1, pch = 16)
points(grd_reg$centres[, 3:4], col = "black", pch = 16, cex = 2)

# Greedy clustering using regular graph
grd_grid <- greedy_cluster(
 x,
 k = 3,
 control = list(
    rng_start = TRUE,
    greedy = 2L,
    neighbours = 1L,
    neighbours_type = "grid",
    seed = 123456
 )
)

plot(x[, 3], x[, 4], col = cl + 1, pch = 16)
points(grd_grid$centres[, 3:4], col = "black", pch = 16, cex = 2)

}
\seealso{
Other greedy cluster:
\code{\link[=control_greedy_cluster]{control_greedy_cluster()}},
\code{\link[=predict.cluster]{predict.cluster()}}
}
\concept{greedy cluster}
