% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_matrix.R
\name{neighbourhood_matrix}
\alias{neighbourhood_matrix}
\title{Create neighbourhood matrix}
\usage{
neighbourhood_matrix(adj_list)
}
\arguments{
\item{adj_list}{A list containing the neighbours of each cluster.}
}
\value{
A neighbourhood matrix.
}
\description{
Create a neighbourhood matrix from an adjacency list.
}
