% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_cluster.R
\name{predict.cluster}
\alias{predict.cluster}
\title{Cluster predictions}
\usage{
\method{predict}{cluster}(object, ...)
}
\arguments{
\item{object}{An object of class \link{cluster}.}

\item{...}{Additional arguments.}
}
\value{
A matrix of allocated clusters.
}
\description{
The allocations predicted by the \link{cluster} object.
}
\details{
A
}
